+++ CHANGELOG.md
@@ -8,7 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ### Added
 
-- Add `mime_override` config field to modify response `Content-Type` [#721](https://github.com/LucasPickering/slumber/issues/721)
+- Add `mime_overrides` config field to modify response `Content-Type` [#721](https://github.com/LucasPickering/slumber/issues/721)
 
 <!-- ANCHOR: changelog -->